\name{aCRM-package}
\alias{aCRM-package}
\alias{aCRM}
\docType{package}
\title{
Convenience functions for analytical Customer Relationship Managment
}
\description{
This package provides functions that are often used in aCRM such as missing values imputation, aggregation and dummy creation, and ensemble modeling.
}
\details{
\tabular{ll}{
Package: \tab aCRM\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2013-03-24\cr
License: \tab GPL (>= 2)\cr
}
Currently provided functions are:
\code{imputeMissings}, \code{Aggregate}, \code{cocktailEnsemble}, \code{predict.cocktailEnsemble}
}
\author{
Dirk Van den Poel, Michel Ballings,  Andrey Volkov, Jeroen D''haen, Michiel Vanherwegen

Maintainer: Michel Ballings  <Michel.Ballings@UGent.be>
}

\references{
Van den Poel, D., Ballings, M., Volkov, A., D''haen, J., Vanherwegen, M., Predictive Analytics for analytical Customer Relationship Management using SAS, Oracle and R, Springer, Forthcoming.
}

\keyword{ package }
\seealso{

 \code{\link{imputeMissings}}, \code{\link{Aggregate}}, \code{\link{cocktailEnsemble}}, \code{\link{predict.cocktailEnsemble}}
}

